stub VivosNextFlow
